Move io.convert_to_yaml into yaml_lib
