+++ yaml_lib/__init__.py
@@ -1,3 +1,4 @@
 """Library to handle all the YAML related functions"""
 
+from .dump_yaml import convert_to_yaml
 from .load_yaml import parse_yaml, read_yaml
